Separate offering operations into a file of their own.
